Adding validation UserService and creating the LoginUser part

+++ BBL/Services/User/UserService.cs
@@ -1,8 +1,11 @@
-﻿using BBL.Repositories.User;
+﻿using BBL.Repositories;
+using BBL.Repositories.User;
 using BBL.Services.User.Models;
 using DataAccessLayer.Entities;
+using System;
 using System.Collections.Generic;
 using System.Data.Common;
+using System.Text.RegularExpressions;
 
 namespace BBL.Services.User
 {
@@ -14,17 +17,49 @@ namespace BBL.Services.User
         {
 			this.userRepository = userRepository;
         }
-	    
+
+		public bool IsValidEmail(string emailAddress)
+		{
+			var pattern = @"^[a-zA-Z0-9.!#$%&'*+-/=?^_`{|}~]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*$";
+
+			var regex = new Regex(pattern);
+			return regex.IsMatch(emailAddress);
+		}
+
+		private bool IsValidPhone(string phoneNumber)
+		{
+			foreach (char digit in phoneNumber)
+			{
+				if (!char.IsDigit(digit) && !char.IsSymbol('+'))
+				{
+					return false;
+				}
+			}
+
+			return true;
+		}
+
 		public int AddUser(AddUserCommand command)
 		{
+			if (string.IsNullOrEmpty(command.Name))
+				throw new Exception(ErrorHandlerService.ParameterNotValidError("Name"));
+
+			if (IsValidEmail(command.Email) || string.IsNullOrEmpty(command.Email))
+				throw new Exception(ErrorHandlerService.ParameterNotValidError("Email"));
+
+			if (string.IsNullOrEmpty(command.Password))
+				throw new Exception(ErrorHandlerService.ParameterNotValidError("Password"));
+
+			if (userRepository.ExistUserByEmail(command.Email))
+				throw new Exception(ErrorHandlerService.ParameterAlreadyExistsError("Email"));
+
+			if (!IsValidPhone(command.Telephone))
+			{
+				throw new Exception(ErrorHandlerService.ParameterMustHaveError("Phone", new[] { "digits" }));
+			}
+
 			var userEntity = userRepository.AddUser(new UserEntity
 			{
-				//Name = "Julia",
-				//Email = "dd.dd@.com",
-				//Password = "1234",
-				//Login = "crme155",
-				//Telephone = "060594313",
-				//IdUserType = 1
 				Name = command.Name,
 				Email = command.Email,
 				Password = command.Password,
@@ -58,7 +93,14 @@ namespace BBL.Services.User
 
 		public GetUserResult GetUserById(int id) 
 		{
+			if (id <= 0)
+				throw new Exception(ErrorHandlerService.ParameterNotValidError("User Id"));
+
 			var userEntity = userRepository.GetUserById(id);
+
+			if (userEntity == null)
+				throw new Exception(ErrorHandlerService.ParameterNotFoundError("User"));
+
 
 			var result = new GetUserResult()
 			{
@@ -74,9 +116,29 @@ namespace BBL.Services.User
 
 		public void UpdateUser(UpdateUserCommand command)
 		{
+			if (command.Id <= 0)
+				throw new Exception(ErrorHandlerService.ParameterNotValidError("User Id"));
+
+			if (string.IsNullOrEmpty(command.Name))
+				throw new Exception(ErrorHandlerService.ParameterNotValidError("Name"));
+
+			if (IsValidEmail(command.Email) || string.IsNullOrEmpty(command.Email))
+				throw new Exception(ErrorHandlerService.ParameterNotValidError("Email"));
+
+			if (userRepository.ExistUserByEmail(command.Email))
+				throw new Exception(ErrorHandlerService.ParameterAlreadyExistsError("Email"));
+
+			if (!IsValidPhone(command.Telephone))
+			{
+				throw new Exception(ErrorHandlerService.ParameterMustHaveError("Phone", new[] { "digits" }));
+			}
+
 			var userEntity = userRepository.GetUserById(command.Id);
 
-			if(command.Name!= userEntity.Name)
+			if (userEntity == null)
+				throw new Exception(ErrorHandlerService.ParameterNotFoundError("User"));
+
+			if (command.Name!= userEntity.Name)
 			   userEntity.Name = command.Name;
 
 			if (command.Telephone!= userEntity.Telephone)
@@ -94,7 +156,13 @@ namespace BBL.Services.User
 
 		public void DeleteUser(int id) 
 		{
+			if (id <= 0)
+				throw new Exception(ErrorHandlerService.ParameterNotValidError("User Id"));
+
 			var userEntity = userRepository.GetUserById(id);
+
+			if (userEntity == null)
+				throw new Exception(ErrorHandlerService.ParameterNotFoundError("User"));
 
 			userRepository.DeleteUser(userEntity);
 		}
@@ -103,6 +171,10 @@ namespace BBL.Services.User
 		{
 			var userResult = userRepository.LoginUser(Login, password);
 
+			if (userResult == null)
+			{
+				return null;
+			}
 			var result = new GetUserResult()
 			{
